+++ frontend/src/app/admin-page/components/StatusLabel.modules.css
@@ -1,0 +1,3 @@
+.pending {
+  background-color: black;
+}
