+++ frontend/src/app/admin-page/components/StatusLabel.tsx
@@ -1,11 +1,12 @@
 /**
  * admin file
  */
+import styles from "./StatusLabel.modules.css";
 
 export type StatusLabelProps = {
   status: "Need to Review" | "Pending" | "Reviewed";
 };
 
 export function StatusLabel({ status }: StatusLabelProps) {
-  return <div>{status}</div>;
+  return <div className={styles.pending}>{status}</div>;
 }
